requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
ConfigArgParse==1.2.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-celery-beat==2.1.0
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-extensions==3.0.9
django-filter==2.4.0
django-import-export==2.5.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-sortedm2m==3.0.2
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
jdcal==1.4.1
jsonfield==3.1.0
kombu==5.0.2
locust==1.4.1
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.2
mutagen==1.45.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
phonenumberslite==8.12.14
Pillow==8.1.0
prompt-toolkit==3.0.8
psutil==5.8.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.5
pywin32==300
PyYAML==5.4.1
pyzmq==20.0.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
swapper==1.1.2.post1
tablib==3.0.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0